moved test to correct package
